cql_additional_tests.py: add require mark to test_whole_list_conditional
As written in https://github.com/scylladb/scylla-dtest/issues/2995,
there were some changes in the way NULLs inside of lists are handled
in the IF conditions. A dtest started failing because of this.

It's not clear what should be done about it,
we could allow the new behavior, or ban it.
Allowing it is risky, because generally lists shouldn't
contain nulls and we could crash somewhere.
Banning it is risky, beacause the new behavior
is in line with the semantics of NULLs inside the IF condition.

For now let's mark this test as `require`,
until we decide what to do about it.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>