tools/files: get_list_of_sstables: use regular expression to match table directory
`startswith` might match irrelevant directories.
For example, if table_name=='cf' then it will match also
a table named `cf0`.  The general structure of the table directory
is `<table>-<schema_uuid>`.

This way, a regular expression may be passed as `table_name`
to enable matching multiple tables in one call, if so desired.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>