cql_additional_tests: update the expected error
we changed the way how we validate the timeuuid when evaluating
the value returned by writetime(). before
https://github.com/scylladb/scylladb/commit/b7bbcdd178088c9fdfd217e9a419821d3a87c345,
the validation of type of parameter was performed by the function itself
in make_min_timeuuid_fct() and make_max_timeuuid_fct(). but after the
change, it is performed by prepare_column_mutation_attribute(), which
throws invalid_request_exception with a different error message.

so, in this change, to be backward compatible, we adapt the test
to accept with the new error type and the new error message.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>