tools: add a TimeUUID class
`timeuuid` is a CQL type which is represented
as a UUID v1 value with some timestamp.

When Cassandra compares two TimeUUID values,
it first looks at the timestamp - if one timestamp
is less than the other, then this value is smaller.

In case the two timestamps are the same,
it takes the last 64 bits, interprets them as
a signed 64-bit integer, and compares those
integers.

The python driver doesn't have any special class
to represent `timeuuid` values, it keeps them
in `uuid.UUID`. The problem is that the comparator
of `uuid.UUID` doesn't compare values by timestamps,
it just compares the bytes lexicographically.

This causes problems - sometimes we would like
to compare two `timeuuid` values in the same
way that Cassandra does, but there's currently
no way to do it.

`class TimeUUID` is a wrapper over `uuid.UUID`,
which has the same comparison operators as
Cassandra's `timeuuid`.

In the future it might be added to the driver:
https://github.com/scylladb/python-driver/issues/231,
but for now we need a custom wrapper to compare TimeUUID
values in the same way that Cassandra does.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>