fix(cql_group_functions_tests.py): compare timeuuids correctly
This test used `uuid.UUID` to represent values of type `timeuuid`.
It tests the aggregate functions `max()` and `min()`, which choose
the largest/smallest `timeuuid`.

The problem is that `uuid.UUID` comparator works differently
from the one that Cassandra uses for `timeuuid` values.

`timeuuid` values should be compared first by the timestamp,
and then based on the remaining bytes. `uuid.UUID` compares
the bytes lexicographically, which isn't the same.

For example, Cassandra says that `fed35080-0efb-11ee-a1ca-00006490e9a4`
is smaller than `00000257-0efc-11ee-9547-00006490e9a6`,
but `uuid.UUID` comparator says the opposite - it thinks that the one
that starts with zeros is smaller.

To fix this, let's sort `timeuuid` values using `TimeUUID` comparator,
instead of the `uuid.UUID` one.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>