audit_tests: make code more consistent
This patch unifies the audit log querying and testing utilities.
1. It adds a central function that query the audit log, sorts it
   properly and return the log as a list.
2. It changes the validation function to be field based so it is clearer
   what the comparison is doing.
3. I adds an assert for the field order in the log, this is only in
   order to preserve a failure that would have happened if order of
   fields doesn't align with our expectation (meaning it has changed).

NOTE: There is an additional change that has been made to support the
above, this change is forcing the row_factory for the session to be
named tuple factory, so we can refer to different log fields by name
instead of by offset.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>