audit_tests: fix test_audit_audit_ks
The test assumed that the log is formed before the query,
this is untrue since we also log success state which means that we first
need to run the query and only then log it.
The fix is eventually very simple and involves making the query before
we test the log.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>