pipelines: stop failing on reaching splitMaxNodes
change of the semantic of how `splitMaxNodes` is used.
till new if the split was reaching this number, the build would be failing
stright away.
now it would run only the first splits reaching `splitMaxNodes`
while issueing a warning that would be written on the job description.

this mean that when reaching this limit, the longest tests wouldn't be running.
(since they are the last to be selected by the splitter logic)