fix(cql_additional_tests.py): adjust tests to allow nulls in LWT lists
As written in #2995, there were some changes in the way NULLs
inside of lists are handled in the IF conditions.
Some dtests started failing because of this.

Let's modify the dtests so that they pass on the new
Scylla which handles NULLs a bit differently.

NULL in lists used to be disallowed.
Now it's allowed, but only inside
the IF conditions.

Refs: https://github.com/scylladb/scylla-dtest/issues/2995

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>