cql_prepared_test.py: add a require marker to test_lwt_update_prepared
This test runs the following queries

```cql
INSERT INTO boolean_update_test_table (k, value) VALUES (2, null)
UPDATE boolean_update_test_table SET value=:upd_v WHERE k=2 IF value in (null);
```

The LWT update should be applied, because `value IS NULL`, so `IF value in (null)` is True.

The problem is that the test passes the value as a bound variable,
and there's currently a python driver bug which causes bound lists
with NULL values to behave differently than they should.
The driver sends a with `empty` values instead of `nulls`.
https://github.com/scylladb/python-driver/issues/201

For now let's mark this test as `require`, so that it doesn't
show up in the failed dtests list.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>