fix(cql_prepared_test.py): skip test cases with list of NULLs
Skip the test cases that test 'value IN :v` with `:v = [null]`.

There's a bug in the python driver which causes them to fail.
Because of the bug the driver sends a list of [empty] instead of [null],
and the LWT update isn't applied.

The `require` mark can be removed now.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>