consistent_cluster_management: use force_schema_commit_log
The problem was with the test
upgrade_test.TestUpgradeWithExperimentalRaft.test_upgrade_cluster_nodes_with_twcs,
it failed with the message on node log:
"Bad configuration: consistent_cluster_management
requires schema commit log to be enabled".

The problem is that schema commitlog must be
enabled before enabling raft. When existing user
is migrating to raft, they need to restart the
cluster with new code without raft and wait for
schema commitlog feature to spread around the cluster.
Then another restart with enabled raft.
For new users there is a force_schema_commit_log
flag in a config which allows to use schema commit
log without waiting for the feature.

There's already a full-fledged upgrade test for
this scenario in test_raft_upgrade_no_schema.py,
so here we just make the existing dtests pass
with the help of force_schema_commit_log.

Fixes: scylladb/scylladb#14176