compaction_additional_test: Enable test_enable_disable_optimized_query_for_twcs on debug mode
test_enable_disable_optimized_query_for_twcs had to be disabled for
debug build, due to timeouts.

The reason for it taking too long is high write amplification as
a result of flushing too often, once every 15s in the simulation.

Test time was reduced from ~30 min to ~7 min in debug mode.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>