treewide: add tests specifically testing java tools to next_gating
quite a few tests using the java tools for accessing the sstable
failed when we are switching from the integer based SSTable
identifier/generation to the UUID-based identifier.

in hope to identifiy the issue, and prevent it from happending again
in future, let's add some of them to next_gating.

Fixes #3274
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>