test(materialized views): range tombstones
Reproducer for issue https://github.com/scylladb/scylla-enterprise/issues/3072#issuecomment-
1605647790:

When a base table of a materialized view is updated, the affected rows are also changed in the
materialized view. For DELETE statements, many rows can be affected, and so the view update
code splits the work into batches. This split was not performed correctly when range tombstones
were involved.
When the view_updating_consumer exceeds its buffer size limit, it flushes the mutation fragment
stream in the middle of a partition. But it doesn't take care to maintain range tombstones
properly while doing this, and if the buffer limit is exceeded in the middle of a range tombstone,
the mutation fragment stream will end with an unclosed range tombstone, which is illegal.