Speed up cluster boot using Scylla options
Today, dtest starts a single node fairly quickly, but starting a multi-
node cluster is painfully slow - for example, test_boot_three_nodes
(which starts a 3-node cluster) takes 119 seconds.

It turns out that the main reason for this slowness is that Scylla uses
sleeps of constant and completely arbitrary durations - 5 or 10 seconds -
waiting for "gossip to settle" when changing the cluster topology. But
dtest runs inside a single machine so has a 100% reliable and very low
latency network, and also does node operations one at a time, so all these
long sleeps are unnecessary. In this patch we use the configuration options

        skip_wait_for_gossip_to_settle: 0

To eliminate these unnecessary sleeps.

With this patch the 119-second run time of cluster with 3 nodes is down
to 15 seconds. If we measure only the cluster boot time and not overheads,
the ratio is even more impressive - from 115 seconds before this patch to
11 seconds.

These numbers can be reduced further later (4 seconds to boot every node
is still sub-optimal), but it's a very good start.

Fixes scylladb#2429

This also add new marker named `no_boot_speedups` that can be used
to to disable this speedup, for those test that it currently
breaking them, until we'll figure out for each one, what
is the reason and how it can be handled.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>