audit_test: introduce an execute_and_validate_audit_entry method
Fragments such as:
    session.execute("USE ks")
    self.assertLastAuditRow(session, "DML", 'USE "ks"',
                            match='DML' in audit_settings['audit_categories'])
Appear in verify_keyspace and verify_table methods.

This way of testing is quite error-prone - there is no check if the last
row was created by most recently executed query. Semantic of the match
argument is weak, is not possible to determine if the last query did not
generate any audit entry. There is even a bug in assertLastAuditRow -
len(res_list) always has to be > 0, even when match=False, so if the
audit log is empty, every call to assertLastAuditRow(..., match=False)
will fail...

execute_and_validate_audit_entry method was created to make testing more
rigorous. By transforming the above fragment of code to:
    self.execute_and_validate_audit_entry(session, "USE \"ks\"", "DML",
                                          audit_settings)