scylla_sstable_test: add support for two representation types for timestamps
this change is a follow-up of 8b9f15f62e491d64a16dd1fbcb7dca90835a553e.
it serves the same purpose -- to ensure that the existing test still
passes before and after scylladb/scylladb#14726 gets merged. this change
fixes the formatting of timestamps so it conforms to ISO 8601 which
requires that the formatted timestamp includes the timezone at the end
of it. after the change in scylladb is merged, we can drop the
"add_z" workaround and the accompanying comment.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>