commitlog_test.py: add test reproducing data resurrection via pinned commitlog segment
Thist commit adds a reproducer for a data resurrection issue that was
thought to be theoretically possible but wasn't confirmed. This test
provides an executable reproducer, confirming that the issue indeed does
exist.
Since the test is failing, it is marked xfail, allowing us to validate a
later fix.

The scenario is as follows:
* Two tables, ks1.tbl1 and ks2.tbl2.
* A commitlog segment contains writes for both tables.
* ks1.tbl1 has very low write traffix, memtables are very rarely
  flushed, this results in ks1.tbl1 "pinning" any commitlog segment
  which has writes, that are still in the memtable.
* ks2.tbl2 deletes the data, some of which is still in the segment
  pinned by ks1.tbl1.
* ks2.tbl2 flushes the memtable, data gets to sstables, commitlog
  segments containing writes for both tables are still pinned.
* ks2.tbl2 compacts and purges away both the data and the tombstone.
* The node has an unclean restart, the pinned segments are replayed and
  data is resurrected, as the tombstone is already purged.

This test uses gc_grace_seconds=0 to make the test fast. With a normal
10 days long gc_grace_seconds, this bug is very unlikely to reproduce,
as it would require a memtable not being flushed for 10 days. But with
lower gc periods, or repair based tombstone gc, the bug is more likely
to be hit.

Refs: https://github.com/scylladb/scylladb/issues/14870