rm json2sstable.py
There is a single test in this file: it creates a cluster, writes some
data to it, dumps the sstables to json, destroys the cluster, converts
json to sstables, then imports the sstables and verifies the result.
One component of this test, json2sstable was deprecated in Cassandra as
of 3.0, see https://issues.apache.org/jira/browse/CASSANDRA-9618.
We should remove all usages of this now deprecated tool, but removing it
from this test makes the test pointless. We could rewrite the test such
that it uses scylla sstable dump-data + scylla sstable write, instead
of sstabledump + json2sstable, but we already have similar sstable ->
json -> sstable test in scylladb.git/test/clq-pytest/test_tools.py.
So instead we should just drop this test as-is, so we don't depend on
deprecated C* tools that no one uses anymore. Doing so opens up the way
for removal of node.run_json2sstable().