limits_test: increase the MAX_COLUMNS limit
In scylladb/scylladb@b3be961, we increased the default schema
commitlog segment size. This change also increased the maximum
number of columns in a table, which caused
test_max_columns_and_query_parameters to fail.

To fix this, we increase the MAX_COLUMNS limit by 4. However, we
don't change this limit in the Raft mode because the number of
columns is bounded by server::configuration::max_command_size
there, and this constant hasn't changed.

Also, increasing MAX_COLUMNS forces the test to run with 4 GB
memory.