json_test.py: adjust expected results for C*'s JSON format
json_test.py was recently changed to also accept timestamps in JSON with
`Z` prefix, as C* (and ISO 8601) requires it.
As we recently discovered, C* strays from ISO 8601 in its JSON
representation of timestamps, in that it replaces the `T` with a space.
This is very strange, but we decided to follow C* in this. Thus, the
expected results in json_test.py need adjusting for this.