conftest: limit `elk_reporter.slices_query_fmt` to last 30 days
currently we query the test duration based on all of it's history
of successful runs.

when we "fix" to make some test run time small, the query still
would return a bigger durtaion (cause of all of the history of that test)
also, in #3320 we are cutting the whole run time in half, so
this is become a limiting factor in our ability to cut on the number
of machines used for running the tests.

in this change we are limiting the query to the last 30 days
only, so now after running a while with changes, we'll start
seeing the benifits as we query less and less of the older tests.