update_cluster_layout_tests: use deterministic pk instead
before this change, we

- specify random primary key with uniform distribution in the
  range of [1..10m]. and
- a clustering key with `cluster: uniform(300)`.
- instruct the insert operation to distribute its partition with
  `fixed(1)`
- run s-c with `-n=100`

in hope that s-c generate a table with 100*300 rows, but sometimes,
we end up with 99*300 rows. the reason is that, sometimes, the 100
batches fails to distribute the pk evenly with 100 different values.
in this case, 99 different values. and 2 of the 100 batches share
the same primary key.

so, in this change, we just generate pk with `seq(1..{nr_partitions})`
to ensure that the specified number of rows are generated without
adding the tolerance factor to the verification.

Fixes #3384
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>