topology_test: add several tests to next_gating
Make following tests of the 'topology_test' module to be gating:
- test_decommission, avg run time is 100seconds
- test_decommissioned_node_cant_rejoin, avg run time is up to 180seconds
- test_remove_node_alive_in_gossip, avg run time is up to 130seconds
- test_removenode_rejected_before_decommision_node,
    avg run time is up to 402seconds

Beacuse those have been stable for last 30 days.

The 'test_remove_node_alive' test had 1 failure during that time period.
So, it is not marked for gating.

The following tests:
- test_movement
- test_move_single_node

Are absent in the ES, so, not marked too.