test_remove_garbage_members_from_group0_after_abort_decommission: decrease heartbeat interval
Commit 6cb4c446e2542f1537edbe479babf02aa496dbab
set nodeops_watchdog_timeout_seconds to be the same
and the watchdog update interval of 10 seconds,
meaning that nodes may abort node operations on
watchdog timeout too aggressively if they miss to process the
heartbeat op on time, which might happen if the test
node is overloaded.

Decrease nodeops_heartbeat_interval_seconds to 1 second
to prevent spurious watchdog timeout aborts.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>