sstable_generation_loading_test: ignore LEAK message in stderr
Cassandra's `cassandra.utils.concurrent.Ref$State` tracks the life
cycles of the objects with a global `ReferenceQueue`, and check if
the object is still referenced by JVM when releasing an object
by looking at the global `ReferenceQueue`, if the object to be
reaped is still being tracked, Cassandra considers it as a leak
and reports it via its logging subsystem with ERROR level.

in our test, we not only check the status code of the `sstableloader`
command line tools, but also check its stdout and stderr outputs.

like other tools provided by Cassandra, `sstableloader` also uses
`cassandra.utils.concurrent.Ref` to track the life cycle of the objects
for identify programming issues.

we've been observing error messages like:

```
ERROR 19:54:41,047 LEAK DETECTED: a reference (org.apache.cassandra.utils.concurrent.Ref$State@5b8ad166) to class org.apache.cassandra.io.sstable.format.SSTableReader$InstanceTidier@737064967:/jenkins/workspace/scylla-enterprise/gating-dtest-release/scylla-enterprise/.dtest/dtest-8ncx5iui/test/node1/data_copy/ks/standard1-b4bf40e033c911ee94f3b85ea0756e50/me-2-big was not released before the reference was garbage collected
```

in our tests. this indicates a bug in sstableloader where it fails
release the reference of an object of object of
`cassandra.io.sstable.format.SSTableReader$InstanceTidier`before
trying to GC it. but it is not fatal in the sense that sstableloader is
able to complete its job, and the status code is zero.

this is not a bug in the scylladb. so in hope to alleviate the pain of
looking into the issue, let's ignore this errror and print it out
in the logging message.

Fixes #3464
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>