alternator_utils: wait for stream in the right place
The utility function prefill_dynamodb_table() waits for the DynamoDB
stream to be enabled, but before this patch did this in the wrong place:
it waited *after* writing the data in the table, but it should wait
*before*. If we don't wait before writing, some of the writes may not
reach the stream - and worse, they may even fail (as noted in
https://github.com/scylladb/scylladb/pull/6901). We saw this happening
in practice in several runs of  Closed alternator_stream_tests.py::
TestAlternatorStreams::test_verify_all_nodes_have_same_stream

So this patch inserts the wait in the middle of prefill_dynamodb_table(),
after creating the table but before filling it with data.

This bug only recently began to make a difference and cause tests to
fail, because until recently we have very long unnecessary sleeps when
dtest boots nodes, and in particular, the wait in prefill_dynamodb_table
became a no-op (after the long delay, there was nothing to wait for)
so it didn't matter if we did it in the right place or not.

Fixes #3498.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>