tools/tables_view_manager: wait_for_view: use ConsistencyLevel.QUORUM
Scylla writes the view status with cl=ONE so it is possible
that the view status will be updated on one node (the one
that builds this view) but on others.

The select query we issue to system_distributed uses
cl=ONE by default so it may miss this status.
We see that in e.g. https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/3144/artifact/logs-full.release.010/dtest-gw3.log
```
09:44:08,282 790     tools.tables_view_manager      DEBUG    tables_view_manager.py:687  | test_do_not_finish_view_building_with_hints: wait_for_view ks.t_by_v: node 7169918e-d2e0-42b1-8127-3e7a8b1c74bc/127.0.46.1 is not done yet
09:44:13,285 790     tools.tables_view_manager      DEBUG    tables_view_manager.py:682  | test_do_not_finish_view_building_with_hints: wait_for_view ks.t_by_v: status=defaultdict(<class 'set'>, {'SUCCESS': {UUID('611c239e-8cfd-41ee-905d-e5213adeceae')}, 'STARTED': {UUID('7169918e-d2e0-42b1-8127-3e7a8b1c74bc'), UUID('b91c9f6f-66d9-40df-b318-1893c63fb32b')}})
09:44:13,285 790     tools.tables_view_manager      DEBUG    tables_view_manager.py:687  | test_do_not_finish_view_building_with_hints: wait_for_view ks.t_by_v: node 7169918e-d2e0-42b1-8127-3e7a8b1c74bc/127.0.46.1 is not done yet
09:44:18,286 790     tools.tables_view_manager      DEBUG    tables_view_manager.py:682  | test_do_not_finish_view_building_with_hints: wait_for_view ks.t_by_v: status=defaultdict(<class 'set'>, {'STARTED': {UUID('611c239e-8cfd-41ee-905d-e5213adeceae'), UUID('b91c9f6f-66d9-40df-b318-1893c63fb32b')}, 'SUCCESS': {UUID('7169918e-d2e0-42b1-8127-3e7a8b1c74bc')}})
...
```

where a node's view build status keeps toggling and we fail
to find all nodes' status as SUCCESS at the same time.

This change uses QUORUM consistency level when possible so
the distributed table results are reconciliated (and read-repair
trigerred if needed).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>