update_cluster_layout_tests: do not count rows using `count()`
the goal of the test is to test the data consistency of tables
with large partition after a new node join the cluster. but there
are at least three variables in the equation:

1. cassandra-stress
   cassandra-stress is a randomized test tool for generating load
   with given profile
2. the column to be counted
   only the non-NULL column is counted
2. count()
   count() function does not imply the data consistency of
   the data.

these factors are not directly related to the data to be verified.

so, in this change, let's drop two of the variables by counting
the rows manually instead of relying on the distribution of
the column whose population is specified by cassandra-stress with
the default one "uniform(1..100B)". but we should not rely on this.
count() is a feature orthogonal to the data consistency, so let's
avoid using it.

Refs #3384
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>