test_disable_tombstone_removal_during_reshard: disable compaction before resharding
To start with, the test ignores the `compaction_strategy` parameter.
It is also sensitive to regular (or rehsape) compaction happening
right after resharding and purging the tomestones, causing it
to observe 0 tombstones after resharding.

See https://github.com/scylladb/scylladb/issues/15191

This change sets the compaction_strategy and disables
compaction before restarting the node with smp=4
so `get_number_of_marked_to_delete` can inspect the sstables
in a quiet environment.  Also, wait for the "Resharded"
log messages rather than the "Resharding" message as the latter
is too early and we may sample the sstables before resharding is
done (especially in debug mode).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>