sstable_generation_loading_test: try all ignore_patterns before giving up
this change is a follow up of 1de820af764564e0de59c5aecdd7ebdb7423e34f,
which was supposed to ignore LEAK error message, but it just gave up
and return `True`, if any of the `ignore_pattern` does not match. but
it should keep trying until exhausting all `ignore_patterns`. this
change just trades the for loop with `any` for better readability, and
also tries harder.

Fixes #3464
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>