alternator_utils: allow retry of failed operations
In alternator_utils.py, we currently set botocore's max_attempts parameter
to 0 instead of the default 5. This means that the DynamoDB API driver
will never retry failed requests, even if they are retryable.

The idea of overriding this default originated in Scylla's test/alternator
in the commit
https://github.com/scylladb/scylladb/commit/0b2cf21932c49ee29bd5afadd37e0a07fa0b900e
it was meant to avoid two problems:

1. Avoid slow failing test - if an operation results in a slow failure,
   we want to report this failure immediately instead of retrying it
   5 times and increasing the failed test's time 5-fold.

2. Avoiding retry on timeout - if the timeout was too low (it defaults to
   60 seconds), it is possible for a a slow query (e.g., CreateTable in
   a debug build) to timeout and be retried even though it was actually
   successful, and the retry would fail.

Neither of these issues are important in dtest: 1. We don't have slow
failures any more (these were usually caused by silly bugs in ancient
versions of Alternator), 2. We increased the timeout from 60 seconds to
300 seconds, so that even the slowest operation on the slowest debug
build finishes by that time.

So in this patch we return the default of max_attempts = 5 - if an
operation returns a retryable error (notable a InternalServerError),
botocore will retry it up to 5 times before giving up.

This is important for issue https://github.com/scylladb/scylladb/issues/15260:
When writing to a table with Alternator Streams enabled, the write may
temporarily fail with InternalServerError if it races with a CDC
"generation" switchover caused by a topology change. This shouldn't
be a problem because InternalServerError is a retryable error, so
botocore will retry the operation - but we need to stop configuring
botocore with max_attempts = 0, or it won't do these retries.

Fixes https://github.com/scylladb/scylladb/issues/15260

Signed-off-by: Nadav Har'El <nyh@scylladb.com>