repair_additional_test: add tests for repairing a single node in the cluster
When a single node is repaired (without the `-pr` option)
it is expected it will sync all the token ranges it participates in,
either as the primary owner, or as a secondary replica.

Add a couple tests that verify that by writing keys
to to other nodes in the cluster and then repairing only one node
and then verifying that all nodes have the data.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>