guardrails: fix dtests broken by Replication Strategy guardrail
As for
`replication_test.TestRestrictionReplicationSimpleStrategy.test_create_and_alter_keyspace_when_value_*`
- we're moving it to cql-pytest in https://github.com/scylladb/scylladb/pull/15399
As for `cqlsh_tests.cqlsh_tests.TestCqlshSmoke.test_select_all_cl_quorum` - with the replication strategy
guardrail in place, we're going to generate a warning when creating a ks
with SimpleStrategy, hence we're not checking if creating such a ks
doesn't return any warnings, but doing just the opposite, i.e. assert
that it indeed generates a warning, which is the expected behavior. BUT
this piece of code has been commented out and can be uncommented only
when the actual change in scylla git repo hits master branch.