materialized_views_test: fix a flaky test
The test
materialized_views_test.py::TestMaterializedViews::test_really_complex_repair
was flaky, as noted in https://github.com/scylladb/scylladb/issues/15314.

The test repeated the same mistake three times: It writes a base table with
CL=ONE and assumes it can read the view with CL=ONE. This is *almost*
true, because the test uses N=RF (both are 5) so view updates are synchronous.
BUT, is not correct if the session is not "exclusive" so the driver can
send different read and write requests to different nodes: If we write to
node A and immediately read from node B, node B might not yet be up-to-date,
and the test fails complaining it can't read what it just wrote.

We can fix this mistake in two ways:
1. Using an "exclusive" connection - it will always use the same node
   as the coordinator, for both reading and writing.
2. Writing with CL=THREE, waiting for all three live nodes to be updated
   (in the first stage of this test, 2 of the 5 nodes were killed).

I used the two fixes in two places. The second fix is more complex, but
I preferred it where we are about to shut down the nodes written because
it guarantees a known state of the test: The test always shuts down the
nodes when all 3 nodes were written, instead of some subset of nodes.

Before this patch, in my local setup this test failed in about 1 of 5
runs. After this patch, I ran the test successfully 50 times without
failure.

Fixes https://github.com/scylladb/scylladb/issues/15314

Signed-off-by: Nadav Har'El <nyh@scylladb.com>