compaction_additional_test: verify expected table distribution
according to
https://docs.python.org/3/library/functools.html#functools.reduce,
functools.reduce() implements the foldl operation on a list. so,
if we have

def is_good(x, y):
  return y >= x * 10

reduce(proc, levels)

where `levels` is [1, 10, 100], what it does is:

1. x_1 = is_good(1, 10)    # x_1 := 10 >= 1 * 10 = True
2. x_2 = is_good(x1, 100)  # x_2 := 100 >= True * 10 = True
3. return x_2

but what we expect is something like:

all(is_good, zip(levels[:-1], levels[1:]))

so we can check if the distribution of the adjacent levels follows
our expectation.

so, in this change,

* instead of using a one-liner, we use a loop to verify the behavior
* and use better variable names for the levels

for better readability. as "x" and "y"

Refs https://github.com/scylladb/scylladb/issues/15405
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>