compaction_additional_test: wait until level distribution converges
in test_lcs_table_promotion_after_stcs_migration, we check the
sstable's distribution on different levels to see if it comply
to the requirement of LCS. but we could check when LCS compaction
is still in the middle of action. so the distribution cannot
represent the ending status of LCS. that's why the test fails
randomly -- we could be looking at an unfinished artwork.

to avoid being disappointed, in this change, we show our patience
by giving the "artist" more time by checking the distribution
for up to 10 times, until the distribution does not change.

Fixes https://github.com/scylladb/scylladb/issues/15405
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>