database_metric_test.py: increase number of reads
The test makes 100 read from a user table, then checks that the
database_total_reads metrics has increased by <= 100 + 10%.
With 100 reads, this 10% was often reached by side-reads done by the
driver. To reduce the chances of this, do 1600 reads, so the reads done
by the driver become less significant. To prevent the runtime of the
test from exploding, do these reads parallelly.

Fixes: scylladb/scylladb#15452