test_cdc_large_values: improve utils::memory_limit_reached error detection
The test triggers the exception quite often and it is supposed to filter
this exception out. However, the regex used to match on the exception
message assumes that the name of the read concurrency semaphore used is
`_read_concurrency_sem`, which might not be true in non-OSS versions.

Fix the issue by making the regex more general.

Refs: scylladb/scylla-enterprise#3485