user_functions_test: do not use Java for UDF in tests
we only support Rust and Lua for UDF lanaguages. so to avoid the noise
in the error message, let's use Lua in the UDF used by the test case.

before this change the actual error is

```
Language 'java' is not supported
```

after this change the actual error is

```
Statement on keyspace user_ks cannot refer to a user type in keyspace ks
```

it matches with the expected error message.

please note, this test passes even without the fix. but the test
performed by `assert_invalid` is way too tolerant. it fails to
match the actual message with the expected one. #3248 was
created to address this. but we need to address all the similar
test failures before enable the check in #3248

Fixes https://github.com/scylladb/scylladb/issues/14324
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>