user_functions_test: use correct column name in test
in 41ced5f652, we migrated to pytest. in this commit, we
also changed the schema of `tab` table in `test_udf_overload_test`
from
`(k text PRIMARY KEY, v int)` to
`(v varchar PRIMARY KEY, i int, t text, a ascii)`.

but we somehow failed to update one of the CQL statements in this
test accordingly. so "k" is still used, even it does not exist
in the new table. that's why we have the error message like

```
Unrecognized name k
```

and it does not match with the expected one:

```
Ambiguous call to function overloaded
```

so, in this change, we change "k" to "t", whose type is the same as
"k". this should correct the error message, so we can verify the
error handling.

Fixes https://github.com/scylladb/scylladb/issues/14325
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>