treewide: escape `matching` param passed to assert_invalid()
the `matching` parameter passed to `assert_invalid()` will be used
as a regular expression to be matched with the error message returned by
the execution of the given CQL query string.

in some tests, we actually passes a regular expression with the special
characters like "(foo|bar)". but under most circumstances, we only
use a regular string, and expect that the error message contains it.

somehow, we fail to escape the special characters used by regular
expression, this will either fail the test or just render the test
way too tolerate. for instance, "ks.test3" would match not only
"ks.test3" but also "ks-test3" despite that "ks-test3" is unexpected
string. not to mention "[Syntax error in CQL query]", it even matches
the single space of " ", this is not even close to what we want.

so, in this change, all the callers of `assert_invalid()` are audited,

* the ones which are used as regular string, but include special
  characters are escaped using `re.escape()`. the escape char of "\"
  is not used for better readability.
* for the consistency, the existing places where the `matching`
  parameters are already escaped are switched to `re.escape()`.

this change prepares us for #3248.

Refs #3248
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>