cdc_test: improve error reporting in cdc_test
Most test cases in cdc_test are based on the following pattern: start to
continuously write to the database, do something with the cluster, then
after some delay stop the write thread and check its results. Currently,
if an exception occurs in the write thread which causes it to fail, it
will be reported only at the end of the test. Because a single test case
can take a relatively long time (measured in minutes), this makes it
very hard to correlate the failure with the events reported in the logs
by the DB.

In order to improve debbugability, make sure to always print the fatal
exception at the moment when it occurs. The error message also reports
how long ago the write has been started.

In addition to that, existing log messages from the worker now also
contain the ID of the worker in which the error occured.

Refs: scylladb/scylladb#15293
Refs: scylladb/scylladb#15357
Refs: scylladb/scylladb#15454