cql_additional_tests: match "allow filtering" in error message
scylla returns error like:

> Only EQ and IN relation are supported on the partition key (unless you
> use the token() function or allow filtering)

please note we are using "allow filtering" in the error message. so
instead of matching "ALLOW FILTERING", let's also match the error
message with the lower case "allow filtering".

as a part of the SELECT CQL statement "ALLOW FILTERING" as case
insensitve. but the error message is a part of the public interface,
it would be better to keep it aligned with our document, which uses
"ALLOW FILTERING". so in this change, we also match with the upper case
variant, so we don't have to change this test when we switch over to
the uppercase error message.

also, "MSG_ALLOW_FILTERING" is not used reused outside of this helper
method, we use a local variable for holding this constant. more readable
this way.

Fixes https://github.com/scylladb/scylladb/issues/14321
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>