cql_additional_tests: do not use MSG_ALLOW_FILTERING in CQL statement
we are using MSG_ALLOW_FILTERING as part of CQL statement, but
as its name indicates, MSG_ALLOW_FILTERING is related to a user-facing
message not a part of CQL statement. and we don't need this flexibility
when composing a CQL statement, so let's replace all
"{MSG_ALLOW_FILTERING}" with "ALLOW FILTERING".

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>