nodetool_additional_test: throttle the rate used by c-s
before this change, both "test_concurrent_drain" and
"test_concurrent_restart" use "concurrent_stress" to put the write
load on scylla nodes, 10 clients are running concurrently when
performing the operation. but we don't throttle the number of
operations per second. when a node in scylla cluster is restarting,
there are good chances that the number of hinting hits the limit
and the coordinator returns OVERLOADED error, after 9 consecutive
retries, cassandra-stress fails the operation and bails out. "9"
is the default value of cassandra-stress's "-error" retries option.
because, if cluster is not able to write the hints to the node
being restarted in time, the chances of completing the retried
requests are rare.

so, in this chance, we throttle the cucurrent request per second
to 5000, this should help with the concurrent test. 5000 is picked
to be consistent with "test_disablebinary_and_disablegossip".

Refs https://github.com/scylladb/scylladb/issues/15192
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>